Changed the result to a double so that it will not crash when dealing with decimals

+++ SimpleCalculator/GUI.cs
@@ -14,7 +14,7 @@ namespace SimpleCalculator
     {
         int value1;
         int value2;
-        int result;
+        double result;
         string function = "";
         public GUI()
         {
